requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
timm>=0.9.10
pandas>=2.1
numpy>=1.25
scikit-learn>=1.3
matplotlib>=3.8
tqdm>=4.66
albumentations>=1.3.1
opencv-python-headless>=4.8
datasets
huggingface_hub
Pillow>=10.0
tabulate
